Add report-list package (#13)
* Add first version of ReportList and related components

* Add emphasis to some text in ReportItemModal

* Correct some typos in README.md

* Fix build for report-list package

Updated to work with @glass/base

* Export ReportListProps

* Make border color match other colors in biologics.

* Improve message when no reports are passed to ReportList

* update to alpha.1 version

* ReportList - Wrap loading spinner in div with padding.

* Add PaginationButtons component to @glass/base

* Add snapshot file for PaginationButtons tests

* Update CustomToggle.spec.tsx to use simulate('click')

* report-list: Remove react-bootstrap from dependencies, update base

Will need to do a final commit to update the proper version numbers
before we merge (so we don't pin an alpha version).

* Remove report-list/README.md

* Remove report-list/yarn.lock

* Add pagination-buttons.scss

Meant to add this MANY commits ago.

* Use .hostNodes() with .find() in enzyme tests

* Remove outdated comment.

* Add tests for ReportList, ReportListItem, and ReportItemModal

* Rename data directory to test_data

* Don't render pagination counts if passed invalid data.

Rendering: "1 - 0 of 0" is pretty worthless.

* PaginationButtons: Add test data for invalid counts

* Bump @glass/base version to 0.2.0-fb-add-report-list.1

* Bump version to @glass/report-list to 0.1.0-fb-add-report-list.1

* Bump @glass/base version in @glass/querygrid to 0.2.0-fb-add-report-list

This also bumps the querygrid version to 0.1.6-fb-add-report-list.1

* Remove className from Media.Body

Apparently this is not supported? Compiled earlier, but I think our
recent changes to the react-bootstrap types library made this a
compiler error, even though it worked as expected before (it added the
class to the default classes).

* Bump version numbers in base, querygrid, and report-list

Also updated lockfile.

* Don't propagate click event on ReportListItem links

* Bump report-list package version to 0.1.0-fb-add-report-list.3

I think I incorrectly pushed a stale build in 0.1.0-fb-add-report-list.2

* Remove target="_blank" from ReportList links.

We don't do this anywhere else in the product, so it doesn't make much
sense to do it in only one area. In general it may be a good idea though
but I am not a designer so it's not my decision.

Also bumped the report-list package version to
0.1.0-fb-add-report-list.4

* Add README to report-list package, update repo README to point to it.

* Remove umd from rollup config in report-list package

* Use runUrl instead of detailsUrl when rendering report list items

* Fix ReportItemModal test.

* Bump @glass/base to v0.2.0

* Bump @glass/omnibox to v0.1.10

Because it now depends on @glass/base v0.2.0

* Bump @glass/querygrid to v0.3.0

Because it now depends on @glass/base v0.2.0 and @glass/omnibox 0.1.10

* Bump @glass/domainproperties to v0.0.9

Because it now depends on @glass/base v0.2.0

* Bump @glass/navigation to v0.1.9

Because it now depends on @glass/base v0.2.0

* Bump @glass/report-list to v0.1.0

Initial release of the component.

* Update lockfile.