Only render one version of QueryGridBar DOM
QueryGridBar was rendering two sets of DOM nodes and hiding one version
with a media query. This was causing us to improperly re-use components
and causing multiple chart modals to render at the same time.

This means we no longer render the omnibox in between the buttons/menu
on the left side and pagination buttons on the right side, instead we
always render the omnibox in its own row.