Fix Issue 38186 (#94)
Change QueryGridModel.getAllColumns so that it uses the columns from getDisplayColumns, and adds any additional columns to that set. We need to do this or we don't get the column metadata from the view, which causes omnibox to drop all filters/sorts/etc. on columns that have a different short caption via a custom view.