Add SampleComparison aware ChartMenu (#128)
This PR addresses a few issues, but primarily it adds a few props to QueryGridPanel to allow consumers to use a Chart Menu that is aware of Sample Comparison reports. Consumers must opt into this via showSampleComparisonReports prop. There are also two new event handlers onReportClicked and onCreateReportClicked so consuming applications can override the default behavior when menu items are clicked.

I also added selectionKey as a getter on QueryGridModel and isSampleLookup to QueryColumn.