Simplify getStateQueryGridModel
We now use the spread operator instead of manually checking if each attribute on initProps is undefined.

Removed IStateModelProps. The name was rather ambiguous and after testing in Biologics it provided no types safety (I could add any additional attributes to an initProps object that I wanted, and the code would still compile).